+++ testsuite/systemtest_long/bintools/Scheduler/tests/CHStone/sha/src/sha.h
@@ -25,13 +25,13 @@
 /* Useful defines & typedefs */
 
 typedef unsigned char BYTE;
-typedef unsigned long LONG;
+typedef unsigned int INT32;
 
 #define SHA_BLOCKSIZE		64
 
-LONG sha_info_digest[5];	/* message digest */
-LONG sha_info_count_lo, sha_info_count_hi;	/* 64-bit bit count */
-LONG sha_info_data[16];
+INT32 sha_info_digest[5];	/* message digest */
+INT32 sha_info_count_lo, sha_info_count_hi;	/* 64-bit bit count */
+INT32 sha_info_data[16];
 
 void sha_init ();
 void sha_update (const BYTE *, int);
